requirements.txt
# ============================================================
# AI-POWERED RESUME SCREENER - ESSENTIAL REQUIREMENTS
# Minimal dependencies for core functionality
# Updated: September 19, 2025
# ============================================================

# Core ML and AI packages
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# PDF processing
PyPDF2>=3.0.1

# Web interface
gradio>=4.7.1

# Data processing
numpy>=1.24.3
pandas>=2.1.3

# Environment configuration
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.8.0